more async resource work

+++ Neo/NeoSource/AssetManager.cpp
@@ -24,7 +24,7 @@ static int GetTime(const string name, const stringlist &extensions, u64& time)
 
 void AssetManager::DeliverAssetDataAsync(AssetType assetType, const string& name, AssetCreateParams* params, const DeliverAssetDataCB& cb)
 {
-	Assert(m_assetTypeInfoMap.contains(assetType), std::format("Cannot create asset: {} - unregistered asset type: {}", name, (int)assetType));
+	Assert(m_assetTypeInfoMap.contains(assetType), std::format("Cannot create asset: {} - unregistered asset type: {}", name, assetType));
 
 	auto assetTypeInfo = m_assetTypeInfoMap[assetType];
 	m_assetTasks.AddTask
@@ -55,7 +55,7 @@ void AssetManager::DeliverAssetDataAsync(AssetType assetType, const string& name
 				// if this is a non-optional source file and it wasn't found, then we can't convert the asset
 				if (srcExt.second && (ext == -1 || srcDateStamp == 0))
 				{
-					Log(std::format("Asset {} [{}] missing src file {}", name, (int)assetType, idx));
+					Log(std::format("Asset {} [{}] missing src file {}", name, assetType, idx));
 					missingSrcFile = true;
 				}
 				idx++;
@@ -82,7 +82,7 @@ void AssetManager::DeliverAssetDataAsync(AssetType assetType, const string& name
 
 				// create from data
 				// this can return nullptr if version is old
-				Log(STR("Create {} From Data: {}", name, (int)assetType));
+				Log(STR("Create {} From Data: {}", name, assetType));
 				AssetData* assetData = assetTypeInfo->m_assetCreateFromData(serializedBlock);
 				if (assetData)
 				{
@@ -118,7 +118,7 @@ void AssetManager::DeliverAssetDataAsync(AssetType assetType, const string& name
 			}
 
 			// now create the AssetData from the src files
-			Log(STR("Create {} From Source: {}", name, (int)assetType));
+			Log(STR("Create {} From Source: {}", name, assetType));
 			AssetData *assetData = assetTypeInfo->m_assetCreateFromSource(srcFileMem, params);
 			assetData->name = name;
 			assetData->type = assetType;
